Add confirm dialogue for each file delete. Add basic logging functionalities

+++ rmim.sh
@@ -4,19 +4,56 @@
 #based on a matched *string* in a given directory.
 #You are MUCH better off just using 'rm' wisely.
 
-#Modify the shell's delimiter for use in remove_files() subshell
+#Modify the shell's delimiter for use in remove_files() subshell, specifically
+#to handle the "find" results easily.
 old_IFS="$IFS"
 IFS="
 "
 
+#Create a ~/log folder if none exists.
+LOG_DIR="/home/`whoami`/log"
+
+if [ ! -d "$LOG_DIR" ]
+then
+    mkdir $LOG_DIR
+fi
+
+#If a rmim logfile exists, back it up once.
+if [ -f ${LOG_DIR}/rmimDeletedFiles.log ]
+then
+    mv ${LOG_DIR}rmimDeletedFiles.log ${LOG_DIR}rmimDeletedFiles.previous.log
+fi
+
 #For each string specified as a script argument,
-#Find all files containing the string in the current directory
+#Find all files containing the string in the current directory.
 remove_files()
 {
   for i in $(find $1 -name "*$2*")
   do
-    rm $i
-    echo "Removed $i"
+    echo -n "Delete this file? $i: "
+    while :
+    do
+        read DELETE_DECISION
+        if [ -z $DELETE_DECISION ]
+        then
+            DELETE_DECISION="yes"
+        fi
+        case $DELETE_DECISION in
+            'y'|'yes')
+              rm $i
+              echo $i >> "${LOG_DIR}/rmimDeletedFiles.log"
+              break
+              ;;
+            'n'|'no')
+              echo "File retained"
+              break
+              ;;
+            *)
+              echo "You must make a choice."
+              echo -n "Delete this file? $i: "
+              ;;  
+        esac
+    done
   done
 }
 
@@ -39,11 +76,6 @@ then
 fi
 
 case $DIR in
-    "$1")
-        echo "Script was called non-interactively, removing files..."
-        remove_files $DIR $SEARCH_STRING
-        ;;
-
     "-h")
         echo "This script removes files for you."
         echo "When run with no arguments, it will walk you through it's actions"
@@ -53,24 +85,33 @@ case $DIR in
         echo "      \"rmim <directory> <matched string>\""
         echo "       Deletes all files that match <string> in <directory>. Default directory is \".\""
         echo "      \"rmim\" Run rmim interactively, with a helpful walkthrough"
-i       ;;
+        ;;
+
+    "$1")
+        echo "Script was called non-interactively, removing files..."
+        remove_files $DIR $SEARCH_STRING
+        ;;
+
 
     "***INT***")    
         echo ""
-        echo "---This script is designed to be run as root.---"
-        echo "--- If you are not root, you will get a whack of permissions errors.---"
-        echo -n "Are you sure you want to run this script? [y/n]"
+        echo "--- This script will remoe files if you have permission. BE CAREFUL. ---"
+        echo "--- This script is designed to be run as root. ---"
+        echo "--- If you are not root, you will get a whack of permissions errors. ---"
+        echo ""
+        echo -n "Are you sure you want to run this script? [y/n] "
 
         while :
         do
           read CONT
           case $CONT in
               "y"|"yes")
-                  echo -n "Which directory shall we operate in?[current directory]:"
+                  echo -n "Which directory shall we operate in?[current directory]: "
                   read DIR
-                  echo -n "Please enter the string to be search for."          
+                  echo -n "Please enter the string to be search for: "
                   read SEARCH_STRING
-                  echo "---Removing all results, please wait...---"
+                  echo "--- Thank you, now removing. Confirmation will be requested for each file..."
+                  sleep 2
 
                   #Determine what the current working directory is
                   #May be switched for "." so it can be added to $PATH
@@ -86,7 +127,7 @@ i       ;;
                   break
                   ;;
               *)
-                  echo "Please say yes or no."
+                  echo "Please say yes or no: "
           esac
         done
         ;;
@@ -94,3 +135,4 @@ esac
 
 #Return the Internat Field Seperator to normal (see top of script)
 IFS=$old_IFS
+echo "Safely completed"
